+++ src/Modules/Admin/Widgets/Navs/TenantNavItem.php
@@ -15,7 +15,7 @@ class TenantNavItem extends NavItem
         $this->icon ??= 'network-wired';
         $this->label ??= Yii::t('tenant', 'TENANT_NAME_PLURAL');
         $this->order ??= 80;
-        $this->roles ??= [Tenant::AUTH_TENANT_CREATE, Tenant::AUTH_TENANT_UPDATE];
+        $this->roles ??= [Tenant::AUTH_TENANT];
         $this->routes =  ['admin/tenant'];
         $this->url ??= ['/admin/tenant/tenant/index'];